+++ src/Network/Network/AzureFirewall/NatRule/NewAzureFirewallNatRuleCommand.cs
@@ -22,8 +22,8 @@ using MNM = Microsoft.Azure.Management.Network.Models;
 
 namespace Microsoft.Azure.Commands.Network
 {
-    [Cmdlet(VerbsCommon.New, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "FirewallNatRule", SupportsShouldProcess = true), OutputType(typeof(PSAzureFirewallNatRule))]
-    public class NewAzureFirewallNatRuleCommand : NetworkBaseCmdlet
+    [Cmdlet(VerbsCommon.New, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "FirewallNatRule", SupportsShouldProcess = true, DefaultParameterSetName = AzureFirewallApplicationRuleParameterSets.Default), OutputType(typeof(PSAzureFirewallNatRule))]
+    public class NewAzureFirewallNatRuleCommand : AzureFirewallBaseCmdlet
     {
         [Parameter(
             Mandatory = true,
@@ -39,6 +39,7 @@ namespace Microsoft.Azure.Commands.Network
 
         [Parameter(
             Mandatory = true,
+            ParameterSetName = AzureFirewallApplicationRuleParameterSets.Default,
             HelpMessage = "The source addresses of the rule")]
         public string[] SourceAddress { get; set; }
 
@@ -129,21 +130,23 @@ namespace Microsoft.Azure.Commands.Network
 
                 ValidateIsSinglePortNotRange(DestinationPort.Single());
                 ValidateIsSinglePortNotRange(TranslatedPort);
+
             }
 
-            var networkRule = new PSAzureFirewallNatRule
+            var natRule = new PSAzureFirewallNatRule
             {
                 Name = this.Name,
                 Description = this.Description,
                 Protocols = this.Protocol?.ToList(),
                 SourceAddresses = this.SourceAddress?.ToList(),
+                SourceIpGroups = this.SourceIpGroup?.ToList(),
                 DestinationAddresses = this.DestinationAddress?.ToList(),
                 DestinationPorts = this.DestinationPort?.ToList(),
                 TranslatedAddress = this.TranslatedAddress,
                 TranslatedFqdn = this.TranslatedFqdn,
                 TranslatedPort = this.TranslatedPort
             };
-            WriteObject(networkRule);
+            WriteObject(natRule);
         }
 
         private void ValidateIsSingleIpNotRange(string ipStr)