+++ src/Network/Network/ChangeLog.md
@@ -23,6 +23,7 @@
 
 ## Version 2.1.0
 * Change `Start-AzVirtualNetworkGatewayConnectionPacketCapture.md` and `Start-AzVirtualnetworkGatewayPacketCapture.md` FilterData option examples.
+* Support for IpGroups in AzureFirewall Application,Nat & Network Rules.
 * Add `PrivateRange` parameter to `AzureFirewall`
 	- Updated cmdlet:
 		- New-AzFirewall